cleaned up and moved tools, moved model answer regex to tools
